Basic project of trying to figure out the default paths for .Net Core
Read the notes at the bottom of _Login.cshtml. This is the smallest I think you can make to create a login page without any addtional pages. Unless its a single page app, I wouldn't recommend doing this. It would mix up methods with HomeController and LoginController. Bad coding practice as well. It is not modular as well.